requirements.txt
absl-py==1.1.0
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
blinker==1.4
Brotli==1.0.9
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==36.0.2
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
docstring-parser==0.14.1
dynaconf==3.1.8
efficientnet-pytorch==0.6.3
entrypoints==0.4
executing==0.8.3
Flask==2.0.3
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.46.3
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.3
importlib-metadata==4.11.4
ipykernel==6.13.0
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonargparse==4.9.0
jupyter-client==7.3.1
jupyter-core==4.10.0
kaitaistruct==0.9
kiwisolver==1.4.2
kornia==0.6.5
ldap3==2.9.1
lightning-bolts==0.5.0
lightning-flash==0.7.5
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mitmproxy==8.0.0
msgpack==1.0.3
multidict==6.0.2
munch==2.5.0
nest-asyncio==1.5.5
numpy==1.22.4
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
pretrainedmodels==0.7.4
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
publicsuffix2==2.20191221
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.2
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
pystiche==1.0.1
python-dateutil==2.8.2
pytorch-lightning==1.6.4
pytz==2022.1
PyYAML==6.0
pyzmq==23.0.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.1
scipy==1.8.1
segmentation-models-pytorch==0.2.1
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.36
stack-data==0.2.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
timm==0.5.4
torch==1.11.0+cu113
torch-tb-profiler==0.4.0
torchaudio==0.11.0+cu113
torchmetrics==0.9.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.2.1.post0
typing-extensions==4.2.0
urllib3==1.26.9
urwid==2.1.2
wcwidth==0.2.5
Werkzeug==2.1.2
wsproto==1.1.0
yarl==1.7.2
zipp==3.8.0
zstandard==0.17.0
